Implementazione Delete Per tavolo

+++ src/main/java/it/prova/pokeronline/service/TavoloServiceImpl.java
@@ -8,6 +8,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import it.prova.pokeronline.model.Tavolo;
 import it.prova.pokeronline.repository.TavoloRepository;
+import it.prova.pokeronline.web.api.exception.GiocatoriPresentiException;
+import it.prova.pokeronline.web.api.exception.TavoloNotFoundException;
 
 @Service
 public class TavoloServiceImpl implements TavoloService {
@@ -25,28 +27,68 @@ public class TavoloServiceImpl implements TavoloService {
 	}
 
 	@Override
+	@Transactional(readOnly = true)
 	public Tavolo caricaSingoloElemento(Long id) {
 		return repository.findById(id).orElse(null);
 	}
 
 	@Override
+	@Transactional(readOnly = true)
 	public Tavolo caricaSingoloElementoEager(Long id) {
 		return repository.findByIdEager(id) ;
 	}
 
 	@Override
-	public Tavolo aggiorna(Tavolo tavoloInstance) {
+	@Transactional
+	public Tavolo aggiorna(Tavolo tavoloInstance, Tavolo tavoloCaricatoDalDB) {
+		if(tavoloCaricatoDalDB.getGiocatori().size() > 0)
+			throw new GiocatoriPresentiException("Impossibile eliminare questo Tavolo, sono ancora presenti dei giocatori al suo interno");
+		
 		return repository.save(tavoloInstance);
 	}
 
 	@Override
+	@Transactional
 	public Tavolo inserisciNuovo(Tavolo tavoloInstance) {
 		return repository.save(tavoloInstance);
 	}
 
 	@Override
+	@Transactional
 	public void rimuovi(Long id) {
+		//Mi carico il tavolo e verifico che i giocatori siano 0
+		Tavolo tavolo = repository.findById(id).orElseThrow(() -> new TavoloNotFoundException("Tavolo con id "+ id + " not found"));
+
+		if(tavolo.getGiocatori().size() > 0)
+			throw new GiocatoriPresentiException("Impossibile eliminare questo Tavolo, sono ancora presenti dei giocatori al suo interno");
+		
 		repository.deleteById(id);
 	}
 
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
 }
